+++ tools/utils.sh
@@ -1039,6 +1039,23 @@ set_hostname() {
     unset hname prefix new_hname
 }
 
+complete_initial_setup() {
+    #
+    #  Depending on if prebuilt or not, either setup final tasks to run
+    #  on first boot or now.
+    #
+    if deploy_state_is_it "$deploy_state_pre_build"; then
+        set_new_etc_profile "$setup_final"
+        msg_1 "Prebuild completed, exiting"
+    else
+        $setup_final
+        msg_1 "Please reboot/restart this app now!"
+        echo "/etc/inittab was changed during the install."
+        echo "In order for this new version to be used, a restart is needed."
+        echo
+    fi
+}
+
 #===============================================================
 #
 #   Main