+++ CHANGES.md
@@ -4,6 +4,7 @@ I will try to keep track of changes between releases here
 
 ## Upcomming release, available in devel branch
 
+- fixed incorrect order of tasks when not prebuilding
 - v0.19.4
 - /usr/local/bin/uptime - better option handling
 - v0.19.3